chore: Remove unused bash aliases and configurations feat: Add Dockerfile and devcontainer configuration refactor: Remove unused GitHub Actions workflow The changes involve removing the unused `.bash_aliases` and `.bashrc` files, as well as the unused `docker-image.yml` GitHub Actions workflow and the `install-base.sh` and `install-ibmcloud.sh` scripts. These files were no longer needed and were cluttering the repository.
In addition, the changes introduce a new Dockerfile and devcontainer configuration in the `.devcontainer` directory. This allows for the creation of a development container with the necessary dependencies for working with IBM Cloud and Terraform. The Dockerfile installs various packages and tools, including Terraform, Ansible, and the IBM Cloud CLI. The devcontainer configuration sets up the development container and specifies the post-start command to be executed.

These changes improve the organization and maintainability of the repository by removing unused files and introducing a standardized development environment